+++ src/main/resources/templates/bar.html
@@ -32,13 +32,10 @@
 <!--        </a>-->
 <!--    </div>-->
 
-    <form action="/index" method="post">
+    <form action="/index" method="get">
         <button class="newButton" >
         </button>
     </form>
-
  </div>
-
 </body>
-
 </html>
