--- src/main/resources/templates/test.html
@@ -1,74 +1,0 @@
-<html lang="en" xmlns:th="http://thymeaLeaf.org">
-
-<head>
-
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Бар</title>
-
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.0/jquery.min.js"></script> <!-- jquery -->
-
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0-beta1/dist/css/bootstrap.min.css" rel="stylesheet"
-          integrity="sha384-0evHe/X+R7YkIZDRvuzKMRqM+OrBnVFBL6DOitfPri4tjfHxaWutUpFmBp4vmVor" crossorigin="anonymous">
-
-    <link type="text/css" rel="stylesheet" href="../static/css/bar.css" th:href="@{css/bar.css}"/>
-    <!-- bootStrap & script -->
-
-    <!-- JS -->
-
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0-beta1/dist/js/bootstrap.bundle.min.js"
-            integrity="sha384-pprn3073KE6tl6bjs2QrFaJGz5/SUsLqktiwsUTF55Jfv3qYSDhgCecCxMW52nD2"
-            crossorigin="anonymous"></script> <!-- bootStrap -->
-
-</head>
-
-<body>
-<!-- registration form !!!!! -->
-
-<!-- Кнопка для открытия модального -->
-<button onclick="document.getElementById('id01').style.display='block'">REGISTRATION</button>
-
-<!-- Модальный (содержит форму регистрации) -->
-<div id="id01" class="modal">
-
-    <form class="modal-content" action="/action_page.php">
-        <div class="container">
-            <h1>Форма регистрации</h1> <span onclick="document.getElementById('id01').style.display='none'"
-                                             class="close" title="Закрыть модальный">BACK</span>
-            <p>Пожалуйста, заполните эту форму, чтобы создать учетную запись.</p>
-            <label for="email"><b>Email</b></label>
-            <select id="email"></select>
-            <input type="text" placeholder="Введите вашу почту" name="email" required>
-
-            <label for="psw"><b>Пароль</b></label>
-            <select id="psw"></select>
-            <input type="password" placeholder="Введите пароль" name="psw" required>
-
-            <label for="psw-repeat"><b>Повторить пароль</b></label>
-            <select id="psw-repeat"></select>
-            <input type="password" placeholder="Повторите пароль" name="psw-repeat" required>
-
-            <label>
-                <input type="checkbox" checked="checked" name="remember" style="margin-bottom:15px"> Запомнить меня
-            </label>
-
-            <p>Создавая учетную запись, вы соглашаетесь с нашим <a href="#" style="color:dodgerblue">Условия и
-                конфиденциальность</a>.</p>
-
-            <div class="clearfix">
-                <button type="submit" class="signup">Регистрация</button>
-            </div>
-        </div>
-    </form>
-</div>
-
-
-
-
-
-
-
-</body>
-
-</html>
